requirements.txt
numpy
matplotlib
jupyter
opencv3
pillow
scikit-learn
scikit-image
scipy
h5py
eventlet
flask-socketio
ffmpeg
imageio=2.1.2
pyqt=4.11.4
tensorflow
keras==2.0.9
requests
tornado
pandas